% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential-abundance.R
\name{add_taxonomy_da}
\alias{add_taxonomy_da}
\title{Add taxonomy to differential abundance results}
\usage{
add_taxonomy_da(ps, list_da_asvs, da_res)
}
\arguments{
\item{ps}{A phyloseq object}

\item{list_da_asvs}{Vector of differentially abundant ASV names}

\item{da_res}{Differential abundance results data frame}
}
\value{
Data frame with taxonomy added
}
\description{
Add taxonomy to differential abundance results
}
\keyword{internal}
