% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{calc_pal}
\alias{calc_pal}
\title{Calculate color palette for groups}
\usage{
calc_pal(ps, group_variable)
}
\arguments{
\item{ps}{A phyloseq object}

\item{group_variable}{Name of grouping variable in sample_data}
}
\value{
Character vector of hex colors
}
\description{
Generates an NPG color palette based on number of groups.
}
