% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{Faiths}
\alias{Faiths}
\title{Calculate Faith's phylogenetic diversity}
\usage{
Faiths(x, tree)
}
\arguments{
\item{x}{Abundance matrix (samples as rows)}

\item{tree}{Phylogenetic tree}
}
\value{
Faith's PD values
}
\description{
Calculate Faith's phylogenetic diversity
}
