% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{format_taxonomy}
\alias{format_taxonomy}
\title{Format and clean taxonomy table}
\usage{
format_taxonomy(ps)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A phyloseq object with formatted taxonomy
}
\description{
Fills missing taxonomic ranks with last known classification and adds
'unknown_' prefix where ranks match parent rank. Creates Highest_classified
column combining ASV ID with lowest classified rank.
}
