% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_top_n}
\alias{get_top_n}
\title{Get top N most abundant taxa}
\usage{
get_top_n(ps, n, level = "species", agg = FALSE)
}
\arguments{
\item{ps}{A phyloseq object}

\item{n}{Number of top taxa to return}

\item{level}{Taxonomic level (default: "species")}

\item{agg}{Aggregate at taxonomic level first (default: FALSE)}
}
\value{
Character vector of top taxa names
}
\description{
Get top N most abundant taxa
}
