% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_top_n_group}
\alias{get_top_n_group}
\title{Get top N most abundant taxa per group}
\usage{
get_top_n_group(ps, n, level = "species", var, group = NULL, agg = FALSE)
}
\arguments{
\item{ps}{A phyloseq object}

\item{n}{Number of top taxa to return}

\item{level}{Taxonomic level (default: "species")}

\item{var}{Grouping variable}

\item{group}{Specific group value to filter}

\item{agg}{Aggregate at taxonomic level first (default: FALSE)}
}
\value{
Character vector of top taxa names
}
\description{
Get top N most abundant taxa per group
}
