% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_phylo}
\alias{load_phylo}
\title{Load data into phyloseq object}
\usage{
load_phylo(asvtab, taxa, mapping, tree = NULL)
}
\arguments{
\item{asvtab}{ASV/OTU abundance matrix}

\item{taxa}{Taxonomy table}

\item{mapping}{Sample metadata}

\item{tree}{Optional phylogenetic tree file path}
}
\value{
A phyloseq object
}
\description{
Load data into phyloseq object
}
\keyword{internal}
