% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{meta_to_df}
\alias{meta_to_df}
\title{Convert sample metadata to data frame}
\usage{
meta_to_df(ps)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A data frame containing sample metadata
}
\description{
Convert sample metadata to data frame
}
