% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{phyloseq_adonis}
\alias{phyloseq_adonis}
\title{PERMANOVA on phyloseq object}
\usage{
phyloseq_adonis(ps, dist_matrix, group_variable, ...)
}
\arguments{
\item{ps}{A phyloseq object}

\item{dist_matrix}{Distance matrix (from calc_betadiv or distance())}

\item{group_variable}{Name of grouping variable in sample_data}

\item{...}{Additional arguments passed to adonis2()}
}
\value{
adonis2 result object
}
\description{
Runs adonis2 (PERMANOVA) on a phyloseq object with specified distance matrix.
}
