% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{phyloseq_betadisper}
\alias{phyloseq_betadisper}
\title{Test for homogeneity of group dispersions}
\usage{
phyloseq_betadisper(ps, dist_matrix, group_variable, ...)
}
\arguments{
\item{ps}{A phyloseq object}

\item{dist_matrix}{Distance matrix}

\item{group_variable}{Name of grouping variable in sample_data}

\item{...}{Additional arguments passed to betadisper()}
}
\value{
ANOVA result for betadisper
}
\description{
Runs betadisper and ANOVA to test for differences in group dispersion.
}
