% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-diversity.R
\name{phyloseq_gunifrac}
\alias{phyloseq_gunifrac}
\title{Calculate Generalized UniFrac distance}
\usage{
phyloseq_gunifrac(ps, asdist = TRUE)
}
\arguments{
\item{ps}{A phyloseq object with phylogenetic tree}

\item{asdist}{Return as dist object (default: TRUE)}
}
\value{
Distance matrix or dist object (alpha = 0.5)
}
\description{
Wrapper for GUniFrac calculation on phyloseq objects.
}
