% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_beta_div}
\alias{plot_beta_div}
\title{Plot beta diversity ordination}
\usage{
plot_beta_div(
  ps,
  dist_matrix,
  ordination,
  group_variable,
  add_ellipse = FALSE,
  cols = NULL
)
}
\arguments{
\item{ps}{A phyloseq object}

\item{dist_matrix}{Distance matrix}

\item{ordination}{Ordination object (from calc_betadiv)}

\item{group_variable}{Grouping variable name}

\item{add_ellipse}{Add confidence ellipses (default: FALSE)}

\item{cols}{Custom color palette}
}
\value{
A ggplot object
}
\description{
Creates an ordination plot with PERMANOVA statistics in caption.
}
