% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_taxonomic_comp}
\alias{plot_taxonomic_comp}
\title{Plot taxonomic composition}
\usage{
plot_taxonomic_comp(
  ps,
  tax_level,
  var,
  ord = NULL,
  n_taxa = 10,
  per_group = FALSE,
  groups = NULL,
  agg = FALSE
)
}
\arguments{
\item{ps}{A phyloseq object}

\item{tax_level}{Taxonomic level for plotting}

\item{var}{Faceting variable}

\item{ord}{Order of facet levels}

\item{n_taxa}{Number of top taxa to show (default: 10)}

\item{per_group}{Get top N per group instead of overall (default: FALSE)}

\item{groups}{Groups for per-group top N selection}

\item{agg}{Aggregate at taxonomic level (default: FALSE)}
}
\value{
A ggplot object
}
\description{
Creates stacked bar plot of taxonomic composition.
}
