% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{ps_to_asvtab}
\alias{ps_to_asvtab}
\title{Convert phyloseq OTU table to data frame}
\usage{
ps_to_asvtab(ps)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A data frame containing the OTU/ASV table
}
\description{
Convert phyloseq OTU table to data frame
}
