% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_tab_delim}
\alias{read_tab_delim}
\title{Read tab-delimited file}
\usage{
read_tab_delim(df)
}
\arguments{
\item{df}{Path to tab-delimited file}
}
\value{
A data frame
}
\description{
Read tab-delimited file
}
\keyword{internal}
