% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{Richness}
\alias{Richness}
\title{Calculate observed richness}
\usage{
Richness(x, detection = 0.5)
}
\arguments{
\item{x}{Numeric vector of abundances}

\item{detection}{Detection threshold (default: 0.5)}
}
\value{
Number of taxa above detection threshold
}
\description{
Calculate observed richness
}
