% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{round_any}
\alias{round_any}
\title{Round to nearest value}
\usage{
round_any(x, accuracy, f = ceiling)
}
\arguments{
\item{x}{Numeric value to round}

\item{accuracy}{Rounding accuracy}

\item{f}{Rounding function (default: ceiling)}
}
\value{
Rounded value
}
\description{
Round to nearest value
}
\keyword{internal}
