% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{Shannon.E}
\alias{Shannon.E}
\title{Calculate Shannon effective diversity}
\usage{
Shannon.E(x)
}
\arguments{
\item{x}{Numeric vector of abundances}
}
\value{
Shannon effective (exponential of Shannon entropy)
}
\description{
Calculate Shannon effective diversity
}
