% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{split_taxonomy}
\alias{split_taxonomy}
\title{Split semicolon-delimited taxonomy string into rank columns}
\usage{
split_taxonomy(asvtab)
}
\arguments{
\item{asvtab}{A data frame with a 'taxonomy' column}
}
\value{
A data frame with taxonomy split into Kingdom through Genus columns
}
\description{
Split semicolon-delimited taxonomy string into rank columns
}
\keyword{internal}
