% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\docType{data}
\name{stacked_bar.palette}
\alias{stacked_bar.palette}
\title{Color palette for stacked bar plots}
\format{
An object of class \code{character} of length 14.
}
\usage{
stacked_bar.palette
}
\description{
A 14-color palette suitable for taxonomic composition plots.
}
\keyword{datasets}
