% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{taxonomy}
\alias{taxonomy}
\title{Extract taxonomy table from phyloseq object}
\usage{
taxonomy(ps)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A data frame containing the taxonomy table
}
\description{
Extract taxonomy table from phyloseq object
}
