% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform abundance data}
\usage{
transform(ps, transform = "mss", offset = 1)
}
\arguments{
\item{ps}{A phyloseq object}

\item{transform}{Transformation method: "mss" (minimum sum scaling),
"relative" (relative abundance), or "clr" (centered log-ratio)}

\item{offset}{Pseudocount offset for CLR transformation (default: 1)}
}
\value{
A phyloseq object with transformed abundances
}
\description{
Normalize phyloseq abundance data using various methods.
}
