% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{xyform}
\alias{xyform}
\title{Create formula from variable names}
\usage{
xyform(y_var, x_vars)
}
\arguments{
\item{y_var}{Response variable name}

\item{x_vars}{Character vector of predictor variable names}
}
\value{
A formula object
}
\description{
Create formula from variable names
}
\keyword{internal}
